Refactor and fixes (Read description)
- Rename all `UI.py`s to `ui.py`
- Fix asset-loading in the `pivot_rotation` example
- Get rid of any post-3.9 style type hints
